Added "extracss" sub to the list. We should come up with a way to let people know what this is.
Changed the "cheesemoo" link at the bottom of the page to a userpage link, to match bj's link.

Updated the script. Now it will ignore stuff that's commented out, and subreddit mods can add a tag to their CSS to have the script ignore certain bits of CSS. This should be helpful for subs that redefine the main sub's emotes or somesuch.